bash: Workaround for bugged startup logic in termux
Bash doesn't load the bashrc for interactive login shells. As work-around
for that, the bashrc is sourced by the user profile. This is both debian default
and my behavior. Termux put the code incorrectly into the *global* profile,
resulting in the bashrc being invoked before the user profile has loaded.
With ENV not being set, the bashrc throws an error.

Work-around until termux/termux-packages#7981 is merged and shipped.