Prevent ibus from fucking with my keyboard layout
Some fucktards upstream considered it a good idea to inject ibus into
every Xorg session, including i3. ibus gets started as dbus service and
is thus not visible as systemd unit and thus cannot be masked. It is
also very difficult to diagnose what caused ibus to be started.

When started, ibus races with setxkbmap to reset the keyboard layout
to the most annoying configuration possible.

Whoever designed this, i hope you get stabbed to death.