argv: Replace with awk-based implementation
The shell version (originally from mirabilos i think) does not work
as expected on busybox systems and i have been unable to find out why.

The awk version is written by myself, so i can fix it myself. I also
took the opportunity to implement opportunistic escaping instead of
quoting, with the intent to keep the output clean from backslashes
and quotes if the data does not require them to be present.