greeter: More compact output, to make it still useful when used from xfce-genmon
xfce-genmon is a panel applet that shows the output of a user-specified command.
It can only show one line, so when i let it run the greeter script, the output
of the greeter script must become unreadable when compressed into one line.

The compression happens because i do echo $(...) from genmon.